refactor: Move WriteConfig to pkg/minikube/config
