+++ cmd/minikube/cmd/config/config_test.go
@@ -20,45 +20,7 @@ import (
 	"bytes"
 	"fmt"
 	"testing"
-
-	"k8s.io/minikube/pkg/minikube/constants"
 )
-
-type configTestCase struct {
-	data   string
-	config map[string]interface{}
-}
-
-var configTestCases = []configTestCase{
-	{
-		data: `{
-    "memory": 2
-}`,
-		config: map[string]interface{}{
-			"memory": 2,
-		},
-	},
-	{
-		data: `{
-    "ReminderWaitPeriodInHours": 99,
-    "cpus": 4,
-    "disk-size": "20g",
-    "log_dir": "/etc/hosts",
-    "show-libmachine-logs": true,
-    "v": 5,
-    "vm-driver": "kvm2"
-}`,
-		config: map[string]interface{}{
-			"vm-driver":                 constants.DriverKvm2,
-			"cpus":                      4,
-			"disk-size":                 "20g",
-			"v":                         5,
-			"show-libmachine-logs":      true,
-			"log_dir":                   "/etc/hosts",
-			"ReminderWaitPeriodInHours": 99,
-		},
-	},
-}
 
 func TestHiddenPrint(t *testing.T) {
 	testCases := []struct {
@@ -88,19 +50,5 @@ func TestHiddenPrint(t *testing.T) {
 		if result != test.TestString {
 			t.Errorf("Result %s not match %s", result, test.TestString)
 		}
-	}
-}
-
-func TestWriteConfig(t *testing.T) {
-	var b bytes.Buffer
-	for _, tt := range configTestCases {
-		err := encode(&b, tt.config)
-		if err != nil {
-			t.Errorf("Error encoding: %v", err)
-		}
-		if b.String() != tt.data {
-			t.Errorf("Did not write config correctly, \n\n expected:\n %+v \n\n actual:\n %+v", tt.data, b.String())
-		}
-		b.Reset()
 	}
 }
