+++ pkg/minikube/config/config_test.go
@@ -46,7 +46,7 @@ var configTestCases = []configTestCase{
     "log_dir": "/etc/hosts",
     "show-libmachine-logs": true,
     "v": 5,
-    "vm-driver": "kvm"
+    "vm-driver": "kvm2"
 }`,
 		config: map[string]interface{}{
 			"vm-driver":                 constants.DriverKvm2,
@@ -139,5 +139,19 @@ func Test_readConfig(t *testing.T) {
 
 	if reflect.DeepEqual(expectedConfig, mkConfig) || err != nil {
 		t.Errorf("Did not read config correctly,\n\n wanted %+v, \n\n got %+v", expectedConfig, mkConfig)
+	}
+}
+
+func TestWriteConfig(t *testing.T) {
+	var b bytes.Buffer
+	for _, tt := range configTestCases {
+		err := encode(&b, tt.config)
+		if err != nil {
+			t.Errorf("Error encoding: %v", err)
+		}
+		if b.String() != tt.data {
+			t.Errorf("Did not write config correctly, \n\n expected:\n %+v \n\n actual:\n %+v", tt.data, b.String())
+		}
+		b.Reset()
 	}
 }
